backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
sqlalchemy==2.0.30
pymysql==1.1.1
cryptography==42.0.8
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
openai==1.30.5
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.3.0
openai-whisper==20231117
torch==2.3.0
torchaudio==2.3.0
aiofiles==23.2.1
python-dotenv==1.0.1
